requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
curlify==2.2.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
